requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pandas
aiosqlite
python-dotenv
pydantic_settings
greenlet